requirements_dev.txt
pytest
pytest-django
tox
coverage
codecov
sphinx
sphinx-autobuild
isort
flake8
python-dateutil
django-extensions